+++ src/main/java/com/github/mavogel/ilias/model/IliasAction.java
@@ -28,7 +28,7 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- *
+ * TODO
  *
  * Created by mavogel on 9/11/16.
  */