+++ src/main/java/com/github/mavogel/ilias/state/states/action/RemoveUsersAction.java
@@ -50,14 +50,15 @@ public class RemoveUsersAction implements ChangeAction {
     public String performAction(final ILIASSoapWebservicePortType endpoint, final UserDataIds userDataIds,
                                 final List<IliasNode> nodes) {
         LOG.info("Removing users from groups");
-        confirm();
-
-        final String sid = userDataIds.getSid();
-        try {
-            IliasUtils.removeAllMembersFromGroups(endpoint, sid, nodes);
-        } catch (IOException | JDOMException e) {
-            LOG.error("Error creating xml parser: " + e.getMessage());
+        if(confirm()) {
+            final String sid = userDataIds.getSid();
+            try {
+                IliasUtils.removeAllMembersFromGroups(endpoint, sid, nodes);
+            } catch (IOException | JDOMException e) {
+                LOG.error("Error creating xml parser: " + e.getMessage());
+            }
         }
+
         return "";
     }